IMPRO-1287: Metadata compliance - error on save if not compliant (#954)
* Changes to make IMPROVER outputs metadata compliant, enabling CLI test data to all be made compliant. Unit tests affected by these changes were also fixed.

* Repurpose test previously for unset unit check to reveal result of enforcement check.

* Changes in response to first review.

* Added cloud base height and lightning per unit area to units.py.

* Correct units for cloud base height.

* Requires full name. Should probably add a hierarchy such that the longest partial match is chosen.

+++ lib/improver/wxcode/weather_symbols.py
@@ -441,7 +441,7 @@ class WeatherSymbols(object):
         threshold_coord = find_threshold_coordinate(cube)
         cube_format = next(cube.slices_over([threshold_coord]))
         symbols = cube_format.copy(data=np.full(cube_format.data.shape, -1,
-                                                dtype=np.int))
+                                                dtype=np.int32))
 
         symbols.remove_coord(threshold_coord)
         symbols = add_wxcode_metadata(symbols)
